refactor: moved collections and related child entities to domain group
