+++ src/main/java/se/ltu/eduo/collection/dto/QuizDto.java
@@ -1,13 +1,14 @@
-package se.ltu.eduo.dto;
+package se.ltu.eduo.collection.dto;
 
 import lombok.Value;
+import se.ltu.eduo.collection.model.Quiz;
 
 import java.io.Serializable;
 import java.time.Instant;
 import java.util.UUID;
 
 /**
- * DTO for {@link se.ltu.eduo.model.collection.Quiz}
+ * DTO for {@link Quiz}
  */
 @Value
 public class QuizDto implements Serializable {
