+++ src/main/java/se/ltu/eduo/collection/model/GenerationFocusArea.java
@@ -1,4 +1,4 @@
-package se.ltu.eduo.model.collection;
+package se.ltu.eduo.collection.model;
 
 public enum GenerationFocusArea {
     ENTIRE_MATERIAL("Entire material"),
