+++ src/main/java/se/ltu/eduo/collection/repository/CollectionRepository.java
@@ -1,7 +1,7 @@
-package se.ltu.eduo.repository;
+package se.ltu.eduo.collection.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import se.ltu.eduo.model.collection.Collection;
+import se.ltu.eduo.collection.model.Collection;
 
 import java.util.List;
 import java.util.UUID;
