+++ src/main/java/se/ltu/eduo/collection/repository/GenerationRepository.java
@@ -1,7 +1,7 @@
-package se.ltu.eduo.repository;
+package se.ltu.eduo.collection.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import se.ltu.eduo.model.collection.Generation;
+import se.ltu.eduo.collection.model.Generation;
 
 import java.util.UUID;
