+++ src/main/java/se/ltu/eduo/collection/repository/GenerationSourceMaterialRepository.java
@@ -1,7 +1,7 @@
-package se.ltu.eduo.repository;
+package se.ltu.eduo.collection.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import se.ltu.eduo.model.collection.GenerationSourceMaterial;
+import se.ltu.eduo.collection.model.GenerationSourceMaterial;
 
 public interface GenerationSourceMaterialRepository extends JpaRepository<GenerationSourceMaterial, GenerationSourceMaterial.GenerationSourceMaterialId> {
 }
