+++ src/main/java/se/ltu/eduo/collection/repository/SourceMaterialRepository.java
@@ -1,7 +1,7 @@
-package se.ltu.eduo.repository;
+package se.ltu.eduo.collection.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import se.ltu.eduo.model.collection.SourceMaterial;
+import se.ltu.eduo.collection.model.SourceMaterial;
 
 import java.util.UUID;
