+++ src/main/java/se/ltu/eduo/collection/request/CreateCollectionRequest.java
@@ -1,4 +1,4 @@
-package se.ltu.eduo.dto.request;
+package se.ltu.eduo.collection.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
