+++ src/main/java/se/ltu/eduo/collection/request/UpdateProjectRequest.java
@@ -1,3 +1,3 @@
-package se.ltu.eduo.dto.request;
+package se.ltu.eduo.collection.request;
 
 public record UpdateProjectRequest(String name) {}
