+++ src/main/java/se/ltu/eduo/collection/request/UpdateQuizRequest.java
@@ -1,3 +1,3 @@
-package se.ltu.eduo.dto.request;
+package se.ltu.eduo.collection.request;
 
 public record UpdateQuizRequest(String name, String rawContent) {}
